Removed distrobox command from Print keybinding
With the current keybinding using distrobox, the keybinding cannot work because distrobox needs of docker to be installed and the user won't be aware about it. Furthermore, distrobox need that the Arch container is pulled and docker daemon will run.
My correction removes distrobox (and all this additional steps) and uses directly `hyprshot`. In this manner, a new user using this hyprland config can directly print the screen without additional configuration.